+++ src/components/profile/InputField.tsx
@@ -3,20 +3,7 @@
 import Image from "next/image";
 import React from "react";
 import profileUploaderIcon from "@/assets/images/profileUploaderIcon.svg";
-
-interface InputFieldProps {
-  name?: string;
-  isImage?: boolean; //이미지input인지
-  isTextArea?: boolean; //textArea인지
-  isServiceType?: boolean; //제공 서비스인지
-  isArea?: boolean; //서비스 가능 지역인지
-  text: string;
-  placeholder?: string;
-  height?: string;
-  error?: string;
-  value?: string | string[];
-  onChange?: (value: string | string[]) => void;
-}
+import { InputFieldProps } from "@/types/profile.types";
 
 //이미지input / textArea / 제공 서비스 / 지역선택    각 형식을 [isprops] === boolean을 통해 분기처리
 function InputField({
